Replace 'int' with 'size_t' so 64-bit build works
git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@1025 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ tools/iMesh/iMesh_MOAB.cpp
@@ -2544,5 +2544,5 @@ void cfptr_(void **instance)
 
 void cfptr2_(void *instance) 
 {
-  std::cout << "Instance ptr = " << (int) instance << std::endl;
+  std::cout << "Instance ptr = " << (size_t) instance << std::endl;
 }
